pyproject.toml
[tool.poetry]
name = "cowtrack-backend-2"
version = "0.1.0"
description = "V2 version of cowtrack"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-money = "^3.4.0"
drf-writable-nested = "^0.7.0"
djoser = "^2.2.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
